Struct celly::grid::twodim::GridCoord

pub struct GridCoord {
    // some fields omitted
}

Coordinate for TwodimGrid. Need for custom coordinate raises from the fact that grid is using one-dimensional Vec to store cells. This coordinate can be constructed from offset and grid size.

Methods

impl GridCoord

fn from_offset(offset: u32, rows: u32, cols: u32) -> GridCoord

Constructs GridCoord from given offset (in one-dimensional array or Vec) and grid size.