Trait celly::traits::Cell

pub trait Cell: Serialize + Deserialize {
    type Coord: Coord;
    type State: EvolutionState;
    fn update<'a, I>(&'a mut self, old: &'a Self, neighbors: I, &Self::State) where I: Iterator<Item=Option<&'a Self>>;
    fn with_coord<C: Coord>(C) -> Self;
    fn coord(&self) -> &Self::Coord;
    fn set_coord<C: Coord>(&mut self, &C);
}

Main trait should be implemented in user's code. Such structs contain main logic of cellular automaton. Grids can handle only one cell, so it should be all-in-one.

Associated Types

type Coord: Coord

Coords supported by Cell.

type State: EvolutionState

Global state of evolution.

Required Methods

fn update<'a, I>(&'a mut self, old: &'a Self, neighbors: I, &Self::State) where I: Iterator<Item=Option<&'a Self>>

This method is called for every instance of Cell in grid. Cell can mutate itself. Grid should pass neighbors, previous version of this Cell and the global state.

fn with_coord<C: Coord>(C) -> Self

Constructs Cell with given coord.

fn coord(&self) -> &Self::Coord

Getter for cell's coordinate.

fn set_coord<C: Coord>(&mut self, &C)

Setter for cell's coordinate.