Trait celly::traits::Coord

pub trait Coord: Clone + Serialize + Deserialize {
    fn from_2d(x: i32, y: i32) -> Self;
    fn x(&self) -> i32;
    fn y(&self) -> i32;

    fn z(&self) -> i32 { ... }
}

Basic coordinate with three components. This crate is used for creating automata on 2D grids for now, so z has default impl.

Required Methods

fn from_2d(x: i32, y: i32) -> Self

Build coord from any other representation of coord.

fn x(&self) -> i32

Returns x component.

fn y(&self) -> i32

Returns y component.

Provided Methods

fn z(&self) -> i32

Returns z component.