Trait celly::traits::Grid

pub trait Grid {
    type Cell: Cell;
    type Coord: Coord;
    fn update(&mut self);
    fn state(&self) -> &Self::Cell::State;
    fn cells(&self) -> &Vec<Self::Cell>;
    fn dimensions(&self) -> Self::Coord;
    fn set_cells(&mut self, Vec<Self::Cell>);
}

Grid stores cells and updates them. Also grid contains global evolution state.

Associated Types

type Cell: Cell

Grid wants to work with them.

type Coord: Coord

Grid knows how to work with them.

Required Methods

fn update(&mut self)

One step in evolution.

fn state(&self) -> &Self::Cell::State

Getter for evolution state.

fn cells(&self) -> &Vec<Self::Cell>

Getter for all cells. It is Vec because Rust does not have abstract return types for now. So custom grids are doomed to use Vecs internally.

fn dimensions(&self) -> Self::Coord

Returns Coord with rows and cols counts of grid (2D). 3D grids would have more dimensions.

fn set_cells(&mut self, Vec<Self::Cell>)

This method gives an ability to change grid externally. It could be done from consumer, for example (consider an app where you reacting to user input), or from engine (consider distributed engine received updates from nodes).